优化：将攻略页面的横幅通知从 PageInfoPost 拆分出来
