+++ docs/.vitepress/theme/layouts/Posts.vue
@@ -4,6 +4,7 @@ import { useRouter } from "vitepress";
 import Head from "@/composables/Head.vue";
 import SidebarLeft from "@/composables/sidebar/SidebarLeft.vue";
 import TopNav from "@/composables/top-nav/TopNav.vue";
+import PostBanner from "@/customized/posts/PostBanner.vue";
 import PageTitle from "@/composables/PageTitle.vue";
 import PageInfoPost from "@/customized/posts/PageInfoPost.vue";
 import Main from "@/composables/Main.vue";
@@ -13,6 +14,7 @@ import NetworkWarning from "@/components/common/NetworkWarning.vue";
 
 const router = useRouter();
 
+let postBannerKey = ref(0);
 let titleKey = ref(0);
 let pageInfoPostKey = ref(0);
 
@@ -22,6 +24,7 @@ const hasPageInfoPost = link => {
 
 watch(() => router.route.data.relativePath, (path) => {
     nextTick(() => {
+        postBannerKey.value++;
         titleKey.value++;
         pageInfoPostKey.value++;
     });
@@ -33,6 +36,7 @@ watch(() => router.route.data.relativePath, (path) => {
     <SidebarLeft />
     <main>
         <TopNav />
+        <PostBanner :Key="postBannerKey" :link="$frontmatter.canonical" v-if="hasPageInfoPost($frontmatter.canonical)" />
         <PageTitle :key="titleKey" v-if="!$frontmatter.customTitle" />
         <PageInfoPost :key="pageInfoPostKey" :link="$frontmatter.canonical" v-if="hasPageInfoPost($frontmatter.canonical)" />
         <Main />
